+++ kroxylicious-api/src/test/java/io/kroxylicious/proxy/filter/FilterInvokersTest.java
@@ -6,7 +6,6 @@
 
 package io.kroxylicious.proxy.filter;
 
-import java.util.List;
 import java.util.concurrent.CompletionStage;
 import java.util.stream.Stream;
 
@@ -17,7 +16,6 @@ import org.apache.kafka.common.message.ResponseHeaderData;
 import org.apache.kafka.common.protocol.ApiKeys;
 import org.apache.kafka.common.protocol.ApiMessage;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
@@ -37,30 +35,13 @@ class FilterInvokersTest {
         assertThat(FilterInvokers.from(invalid)).isNotNull();
     }
 
-    @Test
-    void testCompositeFilter() {
-        MultipleSpecificFilter filterA = new MultipleSpecificFilter();
-        RequestResponseFilter filterB = new RequestResponseFilter();
-        RequestFilter filterC = (apiKey, header, body, filterContext) -> null;
-        ResponseFilter filterD = (apiKey, header, body, filterContext) -> null;
-        List<FilterAndInvoker> from = FilterInvokers.from(
-                (CompositeFilter) () -> List.of((CompositeFilter) () -> List.of(filterA, filterB), (CompositeFilter) () -> List.of(filterC, filterD)));
-        List<Filter> filters = from.stream().map(FilterAndInvoker::filter).toList();
-        assertThat(filters).containsExactly(filterA, filterB, filterC, filterD);
-    }
-
     public static Stream<Filter> invalidFilters() {
         Filter noFilterSubinterfacesImplemented = new Filter() {
 
         };
         return Stream.of(noFilterSubinterfacesImplemented,
                 new SpecificAndRequestFilter(),
-                new SpecificAndResponseFilter(),
-                new SpecificAndCompositeFilter(),
-                new CompositeAndRequestFilter(),
-                new CompositeAndResponseFilter(),
-                new TooDeeplyNestedCompositeFilter(),
-                new SelfReferencingCompositeFilter());
+                new SpecificAndResponseFilter());
     }
 
     public static Stream<Filter> validFilters() {
@@ -71,9 +52,7 @@ class FilterInvokersTest {
                 new MultipleSpecificFilter(),
                 requestFilter,
                 responseFilter,
-                new RequestResponseFilter(),
-                new NoRecursionCompositeFilter(),
-                new SingleRecursionCompositeFilter());
+                new RequestResponseFilter());
     }
 
     static class RequestResponseFilter implements RequestFilter, ResponseFilter {
@@ -86,22 +65,6 @@ class FilterInvokersTest {
         @Override
         public CompletionStage<ResponseFilterResult> onResponse(ApiKeys apiKey, ResponseHeaderData header, ApiMessage response, FilterContext context) {
             return null;
-        }
-    }
-
-    static class NoRecursionCompositeFilter implements CompositeFilter {
-
-        @Override
-        public List<Filter> getFilters() {
-            return List.of(new MultipleSpecificFilter());
-        }
-    }
-
-    static class SingleRecursionCompositeFilter implements CompositeFilter {
-
-        @Override
-        public List<Filter> getFilters() {
-            return List.of((CompositeFilter) () -> List.of(new MultipleSpecificFilter()), (CompositeFilter) () -> List.of(new MultipleSpecificFilter()));
         }
     }
 
@@ -118,64 +81,6 @@ class FilterInvokersTest {
         public CompletionStage<ResponseFilterResult> onApiVersionsResponse(short apiVersion, ResponseHeaderData header, ApiVersionsResponseData response,
                                                                            FilterContext context) {
 
-            return null;
-        }
-    }
-
-    static class CompositeAndRequestFilter implements RequestFilter, CompositeFilter {
-        @Override
-        public List<Filter> getFilters() {
-            return null;
-        }
-
-        @Override
-        public CompletionStage<RequestFilterResult> onRequest(ApiKeys apiKey, RequestHeaderData header, ApiMessage request, FilterContext context) {
-
-            return null;
-        }
-    }
-
-    /**
-     * We do not want to explode Composite Filters forever
-     */
-    static class SelfReferencingCompositeFilter implements CompositeFilter {
-        @Override
-        public List<Filter> getFilters() {
-            return List.of(this);
-        }
-    }
-
-    static class TooDeeplyNestedCompositeFilter implements CompositeFilter {
-        @Override
-        public List<Filter> getFilters() {
-            return List.of((CompositeFilter) () -> List.of((CompositeFilter) () -> List.of((ApiVersionsRequestFilter) (apiVersion, header, request, context) -> null)));
-        }
-    }
-
-    static class CompositeAndResponseFilter implements ResponseFilter, CompositeFilter {
-        @Override
-        public List<Filter> getFilters() {
-            return null;
-        }
-
-        @Override
-        public CompletionStage<ResponseFilterResult> onResponse(ApiKeys apiKey, ResponseHeaderData header, ApiMessage response, FilterContext context) {
-
-            return null;
-        }
-    }
-
-    static class SpecificAndCompositeFilter implements ApiVersionsRequestFilter, CompositeFilter {
-
-        @Override
-        public CompletionStage<RequestFilterResult> onApiVersionsRequest(short apiVersion, RequestHeaderData header, ApiVersionsRequestData request,
-                                                                         FilterContext context) {
-
-            return null;
-        }
-
-        @Override
-        public List<Filter> getFilters() {
             return null;
         }
     }